Moved funcs to library
